to refactor erc20 and simple tx related to into their own file
