+++ lib/generator/generator.go
@@ -4,16 +4,12 @@ import (
 	"context"
 	"fmt"
 	"math/big"
-	"sync"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
 
 	"github.com/0glabs/evmchainbench/lib/account"
-	"github.com/0glabs/evmchainbench/lib/contract_meta_data/erc20"
 	"github.com/0glabs/evmchainbench/lib/store"
 	"github.com/0glabs/evmchainbench/lib/util"
 )
@@ -95,124 +91,6 @@ func NewGenerator(rpcUrl, faucetPrivateKey string, senderCount, txCount int, sho
 	}, nil
 }
 
-func (g *Generator) GenerateSimple() (map[int]types.Transactions, error) {
-	txsMap := make(map[int]types.Transactions)
-
-	if g.ShouldPersist {
-		defer g.Store.PersistPrepareTxs()
-	}
-
-	err := g.prepareSenders()
-	if err != nil {
-		return txsMap, err
-	}
-
-	value := big.NewInt(10000000000000) // 1/100,000 ETH
-
-	var mutex sync.Mutex
-	ch := make(chan error)
-
-	for index, sender := range g.Senders {
-		go func(index int, sender *account.Account) {
-			txs := types.Transactions{}
-			for _, recipient := range g.Recipients {
-				tx, err := GenerateSimpleTransferTx(sender.PrivateKey, recipient, sender.GetNonce(), g.ChainID, g.GasPrice, value, g.EIP1559)
-				if err != nil {
-					ch <- err
-					return
-				}
-				txs = append(txs, tx)
-			}
-
-			mutex.Lock()
-			txsMap[index] = txs
-			mutex.Unlock()
-			ch <- nil
-		}(index, sender)
-	}
-
-	for i := 0; i < len(g.Senders); i++ {
-		msg := <-ch
-		if msg != nil {
-			return txsMap, msg
-		}
-	}
-
-	if g.ShouldPersist {
-		err := g.Store.PersistTxsMap(txsMap)
-		if err != nil {
-			return txsMap, err
-		}
-	}
-
-	return txsMap, nil
-}
-
-func (g *Generator) GenerateERC20() (map[int]types.Transactions, error) {
-	txsMap := make(map[int]types.Transactions)
-
-	contractAddress, err := g.prepareContractERC20()
-	if err != nil {
-		return txsMap, err
-	}
-	contractAddressStr := contractAddress.Hex()
-
-	err = g.prepareSenders()
-	if err != nil {
-		return txsMap, err
-	}
-
-	amount := big.NewInt(1000) // a random small amount
-
-	var mutex sync.Mutex
-	ch := make(chan error)
-
-	for index, sender := range g.Senders {
-		go func(index int, sender *account.Account) {
-			txs := types.Transactions{}
-			for _, recipient := range g.Recipients {
-				tx, err := GenerateContractCallingTx(
-					sender.PrivateKey,
-					contractAddressStr,
-					sender.GetNonce(),
-					g.ChainID,
-					g.GasPrice,
-					erc20.MyTokenABI,
-					"transfer",
-					common.HexToAddress(recipient),
-					amount,
-				)
-				if err != nil {
-					ch <- err
-					return
-				}
-				txs = append(txs, tx)
-			}
-
-			mutex.Lock()
-			txsMap[index] = txs
-			mutex.Unlock()
-			ch <- nil
-		}(index, sender)
-	}
-
-	for i := 0; i < len(g.Senders); i++ {
-		msg := <-ch
-		if msg != nil {
-			return txsMap, msg
-		}
-	}
-
-	if g.ShouldPersist {
-		err := g.Store.PersistTxsMap(txsMap)
-		if err != nil {
-			return txsMap, err
-		}
-	}
-
-	return txsMap, nil
-}
-
 func (g *Generator) prepareSenders() error {
 	client, err := ethclient.Dial(g.RpcUrl)
 	if err != nil {
@@ -252,45 +130,4 @@ func (g *Generator) prepareSenders() error {
 	}
 
 	return nil
-}
-
-func (g *Generator) prepareContractERC20() (common.Address, error) {
-	client, err := ethclient.Dial(g.RpcUrl)
-	if err != nil {
-		return common.Address{}, err
-	}
-	defer client.Close()
-
-	tx, err := GenerateContractCreationTx(
-		g.FaucetAccount.PrivateKey,
-		g.FaucetAccount.GetNonce(),
-		g.ChainID,
-		g.GasPrice,
-		erc20.MyTokenBin,
-		erc20.MyTokenABI,
-		"My Token",
-		"MYTOKEN",
-	)
-	if err != nil {
-		return common.Address{}, err
-	}
-
-	err = client.SendTransaction(context.Background(), tx)
-	if err != nil {
-		return common.Address{}, err
-	}
-
-	ercContractAddress, err := bind.WaitDeployed(context.Background(), client, tx)
-	if err != nil {
-		return common.Address{}, err
-	}
-
-	if g.ShouldPersist {
-		g.Store.AddPrepareTx(tx)
-		if err != nil {
-			return common.Address{}, err
-		}
-	}
-
-	return ercContractAddress, nil
 }
