+++ lib/generator/generator_erc20.go
@@ -1,0 +1,121 @@
+package generator
+
+import (
+	"context"
+	"math/big"
+	"sync"
+
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/ethclient"
+
+	"github.com/0glabs/evmchainbench/lib/account"
+	"github.com/0glabs/evmchainbench/lib/contract_meta_data/erc20"
+)
+
+func (g *Generator) GenerateERC20() (map[int]types.Transactions, error) {
+	txsMap := make(map[int]types.Transactions)
+
+	contractAddress, err := g.prepareContractERC20()
+	if err != nil {
+		return txsMap, err
+	}
+	contractAddressStr := contractAddress.Hex()
+
+	err = g.prepareSenders()
+	if err != nil {
+		return txsMap, err
+	}
+
+	amount := big.NewInt(1000) // a random small amount
+
+	var mutex sync.Mutex
+	ch := make(chan error)
+
+	for index, sender := range g.Senders {
+		go func(index int, sender *account.Account) {
+			txs := types.Transactions{}
+			for _, recipient := range g.Recipients {
+				tx, err := GenerateContractCallingTx(
+					sender.PrivateKey,
+					contractAddressStr,
+					sender.GetNonce(),
+					g.ChainID,
+					g.GasPrice,
+					erc20.MyTokenABI,
+					"transfer",
+					common.HexToAddress(recipient),
+					amount,
+				)
+				if err != nil {
+					ch <- err
+					return
+				}
+				txs = append(txs, tx)
+			}
+
+			mutex.Lock()
+			txsMap[index] = txs
+			mutex.Unlock()
+			ch <- nil
+		}(index, sender)
+	}
+
+	for i := 0; i < len(g.Senders); i++ {
+		msg := <-ch
+		if msg != nil {
+			return txsMap, msg
+		}
+	}
+
+	if g.ShouldPersist {
+		err := g.Store.PersistTxsMap(txsMap)
+		if err != nil {
+			return txsMap, err
+		}
+	}
+
+	return txsMap, nil
+}
+
+func (g *Generator) prepareContractERC20() (common.Address, error) {
+	client, err := ethclient.Dial(g.RpcUrl)
+	if err != nil {
+		return common.Address{}, err
+	}
+	defer client.Close()
+
+	tx, err := GenerateContractCreationTx(
+		g.FaucetAccount.PrivateKey,
+		g.FaucetAccount.GetNonce(),
+		g.ChainID,
+		g.GasPrice,
+		erc20.MyTokenBin,
+		erc20.MyTokenABI,
+		"My Token",
+		"MYTOKEN",
+	)
+	if err != nil {
+		return common.Address{}, err
+	}
+
+	err = client.SendTransaction(context.Background(), tx)
+	if err != nil {
+		return common.Address{}, err
+	}
+
+	ercContractAddress, err := bind.WaitDeployed(context.Background(), client, tx)
+	if err != nil {
+		return common.Address{}, err
+	}
+
+	if g.ShouldPersist {
+		g.Store.AddPrepareTx(tx)
+		if err != nil {
+			return common.Address{}, err
+		}
+	}
+
+	return ercContractAddress, nil
+}
